requirements.txt
# PyTorch (CUDA 12.1) 
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# Core scientific stack 
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4

# Config and plotting
PyYAML==6.0.2
matplotlib==3.8.4
seaborn==0.13.2
